requirements.txt
streamlit==1.44.1
requests==2.32.3
PyYAML==6.0.2
